refactor(api): Remove setInterval stream, refactor api struct
