+++ web/api/rest/routes.js
@@ -1,0 +1,7 @@
+module.exports = (server) => {
+  server.express.put('/test', (req, res) => {
+    // Test REST route
+
+    res.send(true);
+  });
+};
